refactor: extract orb slam tracking flow
